+++ represent/resolve.go
@@ -14,7 +14,7 @@ import (
 //
 // It is the responsibility of the edge spec's type handler to determine what "if an edge
 // already exists" means, as well as whether to overwrite/merge or duplicate the edge in such a case.
-func Resolve(g CoreGraph, mid int, src VertexTuple, d EdgeSpec) (StandardEdge, bool) {
+func Resolve(g CoreGraph, mid uint64, src VertexTuple, d EdgeSpec) (StandardEdge, bool) {
 	switch es := d.(type) {
 	case interpret.EnvLink:
 		return resolveEnvLink(g, mid, src, es)
@@ -45,7 +45,7 @@ func Resolve(g CoreGraph, mid int, src VertexTuple, d EdgeSpec) (StandardEdge, b
 	return StandardEdge{}, false
 }
 
-func resolveEnvLink(g CoreGraph, mid int, src VertexTuple, es interpret.EnvLink) (e StandardEdge, success bool) {
+func resolveEnvLink(g CoreGraph, mid uint64, src VertexTuple, es interpret.EnvLink) (e StandardEdge, success bool) {
 	_, e, success = findEnv(g, src)
 
 	// Whether we find a match or not, have to merge in the EnvLink
@@ -80,7 +80,7 @@ func resolveEnvLink(g CoreGraph, mid int, src VertexTuple, es interpret.EnvLink)
 	return
 }
 
-func resolveDataLink(g CoreGraph, mid int, src VertexTuple, es interpret.DataLink) (e StandardEdge, success bool) {
+func resolveDataLink(g CoreGraph, mid uint64, src VertexTuple, es interpret.DataLink) (e StandardEdge, success bool) {
 	e = StandardEdge{
 		Source: src.id,
 		Props:  ps.NewMap(),
@@ -219,7 +219,7 @@ func resolveDataLink(g CoreGraph, mid int, src VertexTuple, es interpret.DataLin
 	return
 }
 
-func resolveSpecCommit(g CoreGraph, mid int, src VertexTuple, es SpecCommit) (e StandardEdge, success bool) {
+func resolveSpecCommit(g CoreGraph, mid uint64, src VertexTuple, es SpecCommit) (e StandardEdge, success bool) {
 	e = StandardEdge{
 		Source: src.id,
 		Props:  ps.NewMap(),
@@ -252,7 +252,7 @@ func resolveSpecCommit(g CoreGraph, mid int, src VertexTuple, es SpecCommit) (e 
 	return
 }
 
-func resolveSpecGitCommitParent(g CoreGraph, mid int, src VertexTuple, es SpecGitCommitParent) (e StandardEdge, success bool) {
+func resolveSpecGitCommitParent(g CoreGraph, mid uint64, src VertexTuple, es SpecGitCommitParent) (e StandardEdge, success bool) {
 	e = StandardEdge{
 		Source: src.id,
 		Props:  ps.NewMap(),
@@ -277,7 +277,7 @@ func resolveSpecGitCommitParent(g CoreGraph, mid int, src VertexTuple, es SpecGi
 	return
 }
 
-func resolveSpecLocalLogic(g CoreGraph, mid int, src VertexTuple, es SpecLocalLogic) (e StandardEdge, success bool) {
+func resolveSpecLocalLogic(g CoreGraph, mid uint64, src VertexTuple, es SpecLocalLogic) (e StandardEdge, success bool) {
 	e = StandardEdge{
 		Source: src.id,
 		Props:  ps.NewMap(),
@@ -304,7 +304,7 @@ func resolveSpecLocalLogic(g CoreGraph, mid int, src VertexTuple, es SpecLocalLo
 	return
 }
 
-func resolveNetListener(g CoreGraph, mid int, src VertexTuple, es SpecNetListener) (e StandardEdge, success bool) {
+func resolveNetListener(g CoreGraph, mid uint64, src VertexTuple, es SpecNetListener) (e StandardEdge, success bool) {
 	// check for existing edge; this one is quite straightforward
 	re := g.OutWith(src.id, Qbe(EType("listening"), "type", "port", "port", es.Port, "proto", es.Proto))
 	if len(re) == 1 {
@@ -332,7 +332,7 @@ func resolveNetListener(g CoreGraph, mid int, src VertexTuple, es SpecNetListene
 	return
 }
 
-func resolveUnixDomainListener(g CoreGraph, mid int, src VertexTuple, es SpecUnixDomainListener) (e StandardEdge, success bool) {
+func resolveUnixDomainListener(g CoreGraph, mid uint64, src VertexTuple, es SpecUnixDomainListener) (e StandardEdge, success bool) {
 	// check for existing edge; this one is quite straightforward
 	re := g.OutWith(src.id, Qbe(EType("listening"), "type", "unix", "path", es.Path))
 	if len(re) == 1 {
@@ -359,7 +359,7 @@ func resolveUnixDomainListener(g CoreGraph, mid int, src VertexTuple, es SpecUni
 	return
 }
 
-func resolveSpecDatasetHierarchy(g CoreGraph, mid int, src VertexTuple, es SpecDatasetHierarchy) (e StandardEdge, success bool) {
+func resolveSpecDatasetHierarchy(g CoreGraph, mid uint64, src VertexTuple, es SpecDatasetHierarchy) (e StandardEdge, success bool) {
 	e = StandardEdge{
 		Source: src.id,
 		Props:  ps.NewMap(),
@@ -388,7 +388,7 @@ func resolveSpecDatasetHierarchy(g CoreGraph, mid int, src VertexTuple, es SpecD
 	return
 }
 
-func resolveSpecParentDataset(g CoreGraph, mid int, src VertexTuple, es SpecParentDataset) (e StandardEdge, success bool) {
+func resolveSpecParentDataset(g CoreGraph, mid uint64, src VertexTuple, es SpecParentDataset) (e StandardEdge, success bool) {
 	e = StandardEdge{
 		Source: src.id,
 		Props:  ps.NewMap(),
@@ -416,7 +416,7 @@ func resolveSpecParentDataset(g CoreGraph, mid int, src VertexTuple, es SpecPare
 	return
 }
 
-func resolveDataProvenance(g CoreGraph, mid int, src VertexTuple, es interpret.DataProvenance) (e StandardEdge, success bool) {
+func resolveDataProvenance(g CoreGraph, mid uint64, src VertexTuple, es interpret.DataProvenance) (e StandardEdge, success bool) {
 	// FIXME this presents another weird case where "success" is not binary. We *could*
 	// find an already-existing data-provenance edge, but then have some net-addr params
 	// change which cause it to fail to resolve to an environment. If we call that successful,
@@ -458,7 +458,7 @@ func resolveDataProvenance(g CoreGraph, mid int, src VertexTuple, es interpret.D
 	return
 }
 
-func resolveDataAlpha(g CoreGraph, mid int, src VertexTuple, es interpret.DataAlpha) (e StandardEdge, success bool) {
+func resolveDataAlpha(g CoreGraph, mid uint64, src VertexTuple, es interpret.DataAlpha) (e StandardEdge, success bool) {
 	// TODO this makes a loop...are we cool with that?
 	success = true // impossible to fail here
 	e = StandardEdge{
